% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_values.R
\name{absValues}
\alias{absValues}
\title{Calculate the absolute values of each measurement in a column}
\usage{
absValues(data, col, new_col)
}
\arguments{
\item{data}{dataframe containing column of interest}

\item{col}{column containing measurements of interest}

\item{new_col}{name new column in which absolute values will be stored}
}
\value{
column The new column containing the calculated absolute values (numeric)
}
\description{
Creates and stores each measurement's absolute value in a new column in the data set
}
