% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateODBA.R
\name{calculateODBA}
\alias{calculateODBA}
\title{Calculate Overall Dynamic Body Acceleration from raw triaxial acceleration values.}
\usage{
calculateODBA(df, x_col, y_col, z_col)
}
\arguments{
\item{data}{dataframe containing the raw sensor data}

\item{x}{column containing raw acceleration values for the X-axis}

\item{y}{column containing raw acceleration values for the Y-axis}

\item{z}{column containing raw acceleration values for the Z-axis}
}
\value{
ODBA column containing the sum of the absolute values of all three axes
}
\description{
Creates and stores ODBA values as a new column in the data set for later use.
}
