% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_Acc_ODBA_plot.R
\name{createBAOPlot}
\alias{createBAOPlot}
\title{Plot ODBA and dynamic acceleration values over time and categorize by behavior}
\usage{
createBAOPlot(data, x, yX, yY, yZ, yODBA, behavior, dest)
}
\arguments{
\item{data}{dataframe containing columns of interest}

\item{x}{column containing timestamps}

\item{yX}{column containing values from the accelerometer's X-axis}

\item{yY}{column containing values from the accelerometer's Y-axis}

\item{yZ}{column containing values from the accelerometer's Z-axis}

\item{yODBA}{column containing the calculated ODBA values}

\item{behavior}{column containing behavioral annotations}

\item{dest}{destination for the plot to be saved}
}
\value{
the final plot
}
\description{
Creates and saves a plot that visualizes the 3 axes of dynamic acceleration and associated calculated ODBA value labelled by behavior observed by an individual over time
}
