% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDatetime.R
\name{createDatetime}
\alias{createDatetime}
\title{Create and format a datetime column from existing date and time information in the dataset}
\usage{
createDatetime(data, date_col, time_col)
}
\arguments{
\item{data}{name of the dataset containing the date and time columns}

\item{date_col}{name of the column containing dates}

\item{time_col}{name of the column containing times}
}
\value{
the original dataset with an additional column containing the newly formatted datetime
}
\description{
Takes the original date and time information, combines it into a new column, and sets the appropriate format
}
