% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBehavior.R
\name{mapBehavior}
\alias{mapBehavior}
\title{Map the spatial distribution of crocodile behaviors}
\usage{
mapBehavior(data, long, lat, ODBA)
}
\arguments{
\item{data}{name of the dataset containing location and behavioral data}

\item{long}{name of column containing longitutde values}

\item{lat}{name of column containing latitude values}

\item{behavior}{name of column containing assigned behavioral states}
}
\value{
a map showing the spatial distribution of crocodile behaviors
}
\description{
Creates a map from GPS data and colors the points according to corresponding behavior
}
