% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapODBA.R
\name{mapODBA}
\alias{mapODBA}
\title{Map the spatial distribution of crocodile activity}
\usage{
mapODBA(data, long, lat, ODBA)
}
\arguments{
\item{data}{name of the dataset containing location and behavioral data}

\item{long}{name of column containing longitutde values}

\item{lat}{name of column containing latitude values}

\item{ODBA}{name of column containing ODBA values}
}
\value{
a map showing the spatial distribution of crocodile activity
}
\description{
Creates a map from GPS data and colors the path according to corresponding ODBA values
}
