% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ODBA.R
\name{ODBA}
\alias{ODBA}
\title{Calculate the ODBA values of each measurement in a column}
\usage{
ODBA(data, col_one, col_two, col_three, new_col)
}
\arguments{
\item{data}{dataframe containing column of interest}

\item{col_one}{column containing absolute values from accelerometer's x-axis}

\item{col_two}{column containing absolute values from accelerometer's y-axis}

\item{col_three}{column containing absolute values from accelerometer's z-axis}

\item{new_col}{name new column in which ODBA values will be stored}
}
\value{
new The new column containing the calculated ODBA values (numeric)
}
\description{
Creates and stores each measurement's ODBA in a new column in the data set
}
