% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performLM.R
\name{performLM}
\alias{performLM}
\title{Perform a linear model with a categorical predictor (behavior) and a numeric response (ODBA)}
\usage{
performLM(data, response_var, predictor_var)
}
\arguments{
\item{data}{name of the data set containing the variables of interest}

\item{response_var}{identifies this column as the response variable in the model}

\item{predictor_var}{identifies this column as the predictor variable in the model}
}
\value{
returns a summary of the linear model
}
\description{
Uses the variables from your data set to run a linear model and summarizes the results
}
