% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\title{Plot the linear model results showing the nature of the relationship between crocodile behaviors and their mean ODBA, or activity levels}
\usage{
plotModel(data, x_var, y_var)
}
\arguments{
\item{data}{name of the data set containing the variables of interest}

\item{x_var}{response variable (ODBA)}

\item{y_var}{predictor variable (behavior)}
}
\value{
a plot visualizing the results of the linear model including means to see the relationship between activity levels and behavior
}
\description{
Creates a plot showing the distribution of ODBA for each behavior
}
