% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readData}
\alias{readData}
\title{Read in data and define NAs}
\usage{
readData(file, na)
}
\arguments{
\item{file}{name of the .csv file to be read in}

\item{na}{defines na values according to how they are formatted in the raw data and converts them to na}
}
\value{
a dataframe containing the raw data and appropriately labeled na values
}
\description{
Imports a csv file, defines the na's, and saves it as a dataframe for further analysis
}
