% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorData.R
\name{sensorData}
\alias{sensorData}
\title{Read in raw sensor data, select columns of interest, and assign sample ID numbers to each row}
\usage{
sensorData(file, col_1, col_2, col_3, col_4, col_5, col_6, sample_id)
}
\arguments{
\item{file}{name of the .csv file to be read in}

\item{sample_id}{creates a column called "sample ID" which numbers the rows in order}
}
\value{
creates a dataframe containing the raw data and numerically labeled rows
}
\description{
Imports a csv file, saves it as a data.frame with only the columns relevant to the current analysis, and creates a sample ID column for further analysis
}
