% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorData.R
\name{sensorData}
\alias{sensorData}
\title{Read in raw data and assign sample ID numbers to each row}
\usage{
sensorData(file, sample_id)
}
\arguments{
\item{file}{name of the .csv file to be read in}

\item{sample_id}{creates a column called "sample ID" which numbers the rows in order}
}
\value{
creates a dataframe containing the raw data and numerically labeled rows
}
\description{
Imports a csv file, saves it as a dataframe, and creates a sample ID column for further analysis
}
